+++ README.rst
@@ -1,5 +1,5 @@
 ------
-LBpy
+pylabolt
 ------
 
 A single phase 2D lattice Boltzmann solver for fluid flow.
