+++ run.py
@@ -1,4 +1,4 @@
-from LBpy.LBpy import main
+from pylabolt.pylabolt import main
 
 
 if __name__ == '__main__':
