+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = LBpy
-version = 0.1
+name = pylabolt
+version = 0.1.1
 author = Malyadeep Bhattacharya, Mithun M Nair
 author_email = malyadeep1999@gmail.com
 url = https://github.com/Malyadeep/LBpy
@@ -33,7 +33,7 @@ example = README.rst, docs/, tests/
 
 [options.entry_points]
 console_scripts = 
-    LBpy = LBpy.LBpy:main
+    pylabolt = pylabolt.pylabolt:main
 
 [project.urls]
 "Homepage" = https://github.com/Malyadeep/LBpy
